# uas-core/src/main/resources/openapi-core-api.json
{
	"openapi": "3.0.4",
	"info": {
		"title": "UNICORE/X Core services",
		"version": "10.2",
		"license": "BSD"
	},

	"servers": [
		{
			"url": "https://localhost:8080/VENUS/rest/core",
			"description": "Test server on localhost"
		}
	],

	"security": [
		{
			"username_password": [],
			"bearer_token": []
		}
	],

	"components": {
		"securitySchemes": {
			"username_password": {
				"type": "http",
				"scheme": "basic"
			},
			"bearer_token": {
				"type": "http",
				"scheme": "bearer"
			},
			"none": {}
		}
	},

	"paths": {

		"/": {
			"get": {
				"description": "Get server's status and user's properties",
				"parameters": [
					{ "$ref": "./openapi-core-components.json#/components/parameters/user_preferences" },
					{ "$ref": "./openapi-core-components.json#/components/parameters/fields" }
				],
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "./openapi-core-components.json#/components/responses/server_status"
								}
							}
						}
					},
					"403": { "description": "Forbidden"	},
					"500": { "description": "Internal server error"	}
				}
			}
		},

		"/certificate": {
			"get": {
				"description": "Get server's certificate",
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"text/plain": {
								"schema": { "type": "string"}
							}
						}
					},
					"500": { "description": "Internal server error" }
				},
				"security": []
			}
		},

		"/storages": {
			"get": {
				"description": "Get server's status and user's properties",
				"parameters": [
					{ "$ref": "./openapi-core-components.json#/components/parameters/user_preferences" },
					{ "$ref": "./openapi-core-components.json#/components/parameters/tags" }
				],
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"$ref": "./openapi-core-components.json#/components/responses/storages_list"
							}
						}
					},
					"403": { "description": "Forbidden" },
					"500": { "description": "Internal server error" }
				}
			}
		},

		"/storages/{id}": {
			"get": {
				"description": "Get storage properties",
				"parameters": [
					{ "$ref": "./openapi-core-components.json#/components/parameters/user_preferences" },
					{ "$ref": "./openapi-core-components.json#/components/parameters/fields" },
					{ "$ref": "./openapi-core-components.json#/components/parameters/id" }
				],
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"$ref": "./openapi-core-components.json#/components/responses/storage_properties"
							}
						}
					},
					"401": { "description": "Unauthorized" },
					"403": { "description": "Forbidden" },
					"404": { "description": "Resource not found" },
					"500": { "description": "Internal server error" }
				}
			}
		},

		"/storages/{id}/files/{path}": {
			"get": {
				"description": "Get properties of a file/directory on the storage",
				"parameters": [
					{ "$ref": "./openapi-core-components.json#/components/parameters/user_preferences" },
					{ "$ref": "./openapi-core-components.json#/components/parameters/fields" },
					{ "$ref": "./openapi-core-components.json#/components/parameters/id" },
					{ "$ref": "./openapi-core-components.json#/components/parameters/file_path" }
				],
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"$ref": "./openapi-core-components.json#/components/responses/files_properties"
							},
							"application/octet-stream":{
								"type": "string",
								"format": "binary"
							}
						}
					},
					"401": { "description": "Unauthorized" },
					"403": { "description": "Forbidden" },
					"404": { "description": "Resource not found" },
					"500": { "description": "Internal server error" }
				}
			},
			"put": {
				"description": "Modify file or its properties",
				"parameters": [
					{ "$ref": "./openapi-core-components.json#/components/parameters/user_preferences" },
					{ "$ref": "./openapi-core-components.json#/components/parameters/id" },
					{ "$ref": "./openapi-core-components.json#/components/parameters/file_path" }
				],
				"requestBody":{
					"content": {
						"application/json": {
							"type": "object"
						},
						"*/*": {
								"type": "string",
								"format": "binary"
						}
					}
				},
				"responses": {
					"204": {
						"description": "No content"
					},
					"401": { "description": "Unauthorized" },
					"403": { "description": "Forbidden" },
					"404": { "description": "Resource not found" },
					"500": { "description": "Internal server error" }
				}
			}
		},

		"/storages/{id}/transfers": {
			"post":{
				"description": "Create a new server-to-server transfer involving this storage",
				"parameters": [
					{ "$ref": "./openapi-core-components.json#/components/parameters/user_preferences" },
					{ "$ref": "./openapi-core-components.json#/components/parameters/id" }
				],
				"requestBody":{
					"required": "true",
					"content": {
						"application/json": {
							"schema": { "$ref": "./openapi-core-components.json#/components/schemas/transfer_description" }
						}
					}
				},
				"responses": {
					"201":{
						"description": "New transfer was created",
						"headers":{
							"Location": {
								"description": "URL of the newly created transfer"
							}
						}
					},
					"401": { "description": "Unauthorized" },
					"403": { "description": "Forbidden" },
					"500": { "description": "Internal server error" }
				}
			 }
		},

		"/transfers": {
			"get": {
				"description": "Lists server-to-server transfers",
				"parameters": [
					{ "$ref": "./openapi-core-components.json#/components/parameters/tags" }
				],
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"$ref": "./openapi-core-components.json#/components/responses/transfers_list"
							}
						}
					},
					"401": { "description": "Unauthorized" },
					"403": { "description": "Forbidden" },
					"500": { "description": "Internal server error" }
				}
			}
		},

		"/transfers/{id}": {
			"get": {
				"description": "Get server-to-server transfer properties",
				"parameters": [
					{ "$ref": "./openapi-core-components.json#/components/parameters/fields" },
					{ "$ref": "./openapi-core-components.json#/components/parameters/id" }
				],
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"$ref": "./openapi-core-components.json#/components/responses/transfer_properties"
							}
						}
					},
					"401": { "description": "Unauthorized" },
					"403": { "description": "Forbidden" },
					"404": { "description": "Resource not found" },
					"500": { "description": "Internal server error" }
				}
			}
		},

		"/jobs": {
			"get": {
				"description": "Get job list",
				"parameters": [
					{ "$ref": "./openapi-core-components.json#/components/parameters/user_preferences" },
					{ "$ref": "./openapi-core-components.json#/components/parameters/tags" }
				],
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"$ref": "./openapi-core-components.json#/components/responses/jobs_list"
							}
						}
					},
					"401": { "description": "Unauthorized" },
					"403": { "description": "Forbidden" },
					"500": { "description": "Internal server error" }
				}
			},
			"post":{
				"description": "Submit a new job",
				"parameters": [
					{ "$ref": "./openapi-core-components.json#/components/parameters/user_preferences" }
				],
				"requestBody":{
					"required": "true",
					"content": {
						"application/json": {
							"schema": { "$ref": "./openapi-core-components.json#/components/schemas/job_description" }
						}
					}
				},
				"responses": {
					"201":{
						"description": "New job was created",
						"headers":{
							"Location": {
								"description": "URL of the newly created job"
							}
						}
					},
					"401": { "description": "Unauthorized" },
					"403": { "description": "Forbidden" },
					"500": { "description": "Internal server error" }
				}
			 }
		},

		"/jobs/{id}": {
			"get": {
				"description": "Get job properties",
				"parameters": [
					{ "$ref": "./openapi-core-components.json#/components/parameters/user_preferences" },
					{ "$ref": "./openapi-core-components.json#/components/parameters/fields" },
					{ "$ref": "./openapi-core-components.json#/components/parameters/id" }
				],
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"$ref": "./openapi-core-components.json#/components/responses/job_properties"
							}
						}
					},
					"401": { "description": "Unauthorized" },
					"403": { "description": "Forbidden" },
					"404": { "description": "Resource not found" },
					"500": { "description": "Internal server error" }
				}
			},
			"put": {
				"description": "Modify job properties",
				"parameters": [
					{ "$ref": "./openapi-core-components.json#/components/parameters/id" }
				],
				"requestBody":{
					"required": "true",
					"content": {
						"application/json": {
							"schema": { "type": "object",
								"properties":{
									"tags" : { "$ref": "./openapi-core-components.json#/components/schemas/tags" },
									"acl": { "$ref": "./openapi-core-components.json#/components/schemas/acl" }
								}
							}
						}
					}
				},
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"type": "object"
							}
						}
					},
					"401": { "description": "Unauthorized" },
					"403": { "description": "Forbidden" },
					"404": { "description": "Resource not found" },
					"500": { "description": "Internal server error" }
				}
			}
		},

		"/jobs/{id}/details":{
			"get": {
				"description": "Get batch system details for the job",
				"parameters": [
					{ "$ref": "./openapi-core-components.json#/components/parameters/id" }
				],
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"type": "object"
							}
						}
					},
					"401": { "description": "Unauthorized" },
					"403": { "description": "Forbidden" },
					"404": { "description": "Resource not found" },
					"500": { "description": "Internal server error" }
				}
			}
		},

		"/jobs/{id}/submitted":{
			"get": {
				"description": "Get the submitted job description",
				"parameters": [
					{ "$ref": "./openapi-core-components.json#/components/parameters/id" }
				],
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"type": "object"
							}
						}
					},
					"401": { "description": "Unauthorized" },
					"403": { "description": "Forbidden" },
					"404": { "description": "Resource not found" },
					"500": { "description": "Internal server error" }
				}
			}
		},

		"/jobs/{id}/forward-port": {
			"get": {
				"description": "Start port forwarding to the application started by this job",
				"parameters": [
					{ "$ref": "./openapi-core-components.json#/components/parameters/id" },
					{
					  "name": "Upgrade", "in": "header", "required": true,
					  "schema": { "const": "UNICORE-Socket-Forwarding" }
					},
					{
					  "name": "host", "in": "query", "description": "Application host to connect to"
					},
					{
					  "name": "port", "in": "query", "description": "Application port to connect to"
					},
					{
					  "name": "file", "in": "query", "description": "Unix domain socket file to connect to"
					},
					{
					  "name": "loginNode", "in": "query", "description": "Specify login node"
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"type": "object"
							}
						}
					},
					"401": { "description": "Unauthorized" },
					"403": { "description": "Forbidden" },
					"404": { "description": "Resource not found" },
					"500": { "description": "Internal server error" }
				}
			}
		}

	}

}